Add files via upload
For my 4th pattern, I was using the while loop and was able to create a pattern I liked but it was only for one line and I was wondering if there was a way to repeat it for more lines like the for loop does. I also tried to use the while loop in the different options so that you can click through them but it wouldnt work so I commented it out and made a new pattern for the first option. I don't know why the while loop pattern wont show up and flip through like the for loop patterns in the other options.